beta3( dsr test ) quickie

+++ Projects/StoreAppLauncher/System.cpp
@@ -1,10 +1,15 @@
 #include "System.h"
 
 System::System( )
-	:m_uNumArgs( 0 ), m_ulAppStatus( STILL_ACTIVE ), m_ulProcessId( 0 ), m_bUseController( false ), m_bBootExplorer( false ), m_iControllerMode( 0 )
+	:m_uNumArgs( 0 ), m_ulAppStatus( STILL_ACTIVE ), m_ulProcessId( 0 ), m_bUseController( false ), m_bBootExplorer( false ), m_iControllerMode( 0 ),
+	m_hProcess( nullptr ), m_hExplorer( nullptr ), m_bCheckParameterForId( true ), m_bUseDSR( false )
 {
 	m_wsArgs.clear( );
 	m_updater.reset( new Updater( ) );
+
+	m_wsKnownArgs = { L"bp", L"bp1", L"dsr" };
+
+	m_dm.dmSize = sizeof( m_dm );
 }
 
 System::~System( )
@@ -46,26 +51,22 @@ bool System::GetAppId( )
 			m_bUseController = true;
 			m_iControllerMode = 1;
 		}
+		if( !wcscmp( args[ i ], L"dsr" ) )
+		{
+			m_bUseDSR = true;
+		}
 	}
-//	if( m_wsArgs.size( ) > 1 )
-//	{
-//		return true;
-//	}
-//	else
-//	{
-//
-//		if( MessageBox( 0, L"Launch parameter was not found. Do you want to update executables within this directory?", L"WARNING: Missing parameter/Update", MB_ICONINFORMATION | MB_OKCANCEL | MB_DEFBUTTON2 ) == IDOK )
-//		{
-//#ifndef _DEBUG
-//			OpenConsole( );
-//#endif
-//			m_updater->Launch( );
-//#ifndef _DEBUG
-//			FreeConsole( );
-//#endif
-//		}
-//	}
-//	return false;
+	if( m_wsArgs.size( ) > 1 )
+	{
+		for( uint i = 0; i < m_wsKnownArgs.size( ); i++ )
+		{
+			if( !wcscmp( m_wsArgs[ 1 ].c_str( ), m_wsKnownArgs[ i ].c_str( ) ) )
+			{
+				m_bCheckParameterForId = false;
+			}
+		}
+	}
+
 	return true;
 }
 
@@ -120,6 +121,11 @@ bool System::OpenAppById( )
 		assert( SUCCEEDED( hr ) );
 	}
 
+	if( m_bUseDSR )
+	{
+		SwitchResolution( );
+	}
+
 	hr = aam->ActivateApplication( m_wsExeName.c_str( ) , nullptr, AO_NONE, &m_ulProcessId );
 	if( FAILED( hr ) )
 	{
@@ -132,29 +138,34 @@ bool System::OpenAppById( )
 				return true;
 			}
 		}
-		MessageBox( 0, L"Please double-check Application ID", L"ERROR: Application ID is probably not correct", MB_ICONERROR | MB_OK );
-		assert( SUCCEEDED( hr ) );
+		if( m_bCheckParameterForId && m_wsArgs.size( ) > 1 )
+		{
+			for( uint i = 0; i < 5; i++ )
+			{
+				Sleep( 100 );
+				hr = aam->ActivateApplication( m_wsArgs[ 1 ].c_str( ), nullptr, AO_NONE, &m_ulProcessId );
+				if( SUCCEEDED( hr ) )
+				{
+					return true;
+				}
+			}
+		}
+		//MessageBox( 0, L"Please double-check id", L"ERROR: Executable name or parameter is probably not correct", MB_ICONERROR | MB_OK );
+		//assert( SUCCEEDED( hr ) );
 	}
 	else
 	{
 		return true;
 	}
-	hr = aam->ActivateApplication( m_wsArgs[ 1 ].c_str( ), nullptr, AO_NONE, &m_ulProcessId );
-	if( FAILED( hr ) )
+
+	if( MessageBox( 0, L"Launch parameter was not found. Do you want to update executables within this directory?", L"WARNING: Missing parameter/Update", MB_ICONINFORMATION | MB_OKCANCEL | MB_DEFBUTTON2 ) == IDOK )
 	{
-		for( uint i = 0; i < 5; i++ )
-		{
-			Sleep( 100 );
-			hr = aam->ActivateApplication( m_wsArgs[ 1 ].c_str( ), nullptr, AO_NONE, &m_ulProcessId );
-			if( SUCCEEDED( hr ) )
-			{
-				return true;
-			}
-		}
-		MessageBox( 0, L"Please double-check Application ID", L"ERROR: Application ID is probably not correct", MB_ICONERROR | MB_OK );
-		assert( SUCCEEDED( hr ) );
+#ifndef _DEBUG
+		OpenConsole( );
+#endif
+		m_updater->Launch( );
 	}
-	return true;
+	return false;
 }
 
 bool System::IsExplorerRunning( )
@@ -231,6 +242,140 @@ void System::GetExecutableName( )
 	wcout << m_wsExeName.c_str( ) << endl;
 }
 
+bool System::GenerateExecutables( )
+{
+//	wstring blacklist[ ] = {
+//		L"Microsoft.3DBuilder",
+//		L"Microsoft.AAD.BrokerPlugin",
+//		L"Microsoft.AccountsControl",
+//		L"Microsoft.Appconnector",
+//		L"Microsoft.Bing",
+//		L"Microsoft.BioEnrollment",
+//		L"Microsoft.CommsPhone",
+//		L"Microsoft.ConnectivityStore",
+//		L"Microsoft.FreshPaint",
+//		L"Microsoft.Getstarted",
+//		L"Microsoft.LockApp",
+//		L"Microsoft.Messaging",
+//		L"Microsoft.MicrosoftEdge",
+//		L"Microsoft.MicrosoftOfficeHub",
+//		L"Microsoft.NET",
+//		L"Microsoft.Office",
+//		L"Microsoft.People",
+//		L"Microsoft.SkypeApp",
+//		L"Microsoft.VCLibs",
+//		L"Microsoft.WinJS",
+//		L"Microsoft.Windows",
+//		L"microsoft.windows",
+//		L"Microsoft.Xbox",
+//		L"Microsoft.Zune",
+//		L"Windows.",
+//		L"windows.",
+//		L"ASUSWelcome"
+//	};
+//#ifndef _DEBUG
+//	OpenConsole( );
+//#endif // !_DEBUG
+//
+//	HString test;
+//	test.Set( RuntimeClass_Windows_Management_Deployment_PackageManager );
+//
+//	ComPtr<IActivationFactory> af;
+//	RoGetActivationFactory( test.Get( ), __uuidof( IActivationFactory ), (void **)&af );
+//
+//	ComPtr<IInspectable> iinsp;
+//	af->ActivateInstance( &iinsp );
+//
+//	ComPtr<IPackageManager2> pm;
+//	iinsp.As( &pm );
+//
+//	ComPtr<Collections::IIterable<Package*>> iiterable;
+//
+//	pm->FindPackagesWithPackageTypes( PackageTypes::PackageTypes_Main, &iiterable );
+//
+//	ComPtr<Collections::IIterator<Package*>> pIter;
+//	iiterable->First( &pIter );
+//
+//	boolean hasCurrent;
+//
+//	for( pIter->get_HasCurrent( &hasCurrent ); hasCurrent; pIter->MoveNext( &hasCurrent ) )
+//	{
+//		ComPtr<IPackage> ipkg;
+//		pIter->get_Current( &ipkg );
+//
+//		ComPtr<IPackageId> ipkgid;
+//		ipkg->get_Id( &ipkgid );
+//		HString name6;
+//		ipkg->GetRuntimeClassName( name6.GetAddressOf( ) );
+//
+//		HString name;
+//
+//		ipkgid->get_FamilyName( name.GetAddressOf( ) );
+//		wstring temp = name.GetRawBuffer( 0 );
+//		m_wsAppsList.push_back( temp );
+//	}
+//	vector<wstring> testvec;
+//	for( uint i = 0; i < ARRAYSIZE( blacklist ); i++ )
+//	{
+//
+//	}
+//	wcout << endl << "Total unfiltered: " << m_wsAppsList.size( ) << endl;
+//	wcout << "Filtered: " << endl;
+//	for( uint i = 0; i < testvec.size( ); i++ )
+//	{
+//		wcout << testvec[ i ].c_str( ) << endl;
+//	}
+//	wcout << endl << "Total filtered: " << testvec.size( ) <<  endl;
+//	system( "pause" );
+//#ifndef _DEBUG
+//	FreeConsole( );
+//#endif
+	return false;
+}
+
+void System::SwitchResolution( bool restore )
+{
+	if( !restore )
+	{
+		EnumDisplaySettings( 0, ENUM_CURRENT_SETTINGS, &m_dm );
+		DEVMODE dmTemp = m_dm;
+
+		vector<ulong> tempResX;
+		vector<ulong> tempResY;
+
+		ulong maxResX = 0;
+		ulong maxResY = 0;
+
+		for( uint i = 0; EnumDisplaySettings( 0, i, &dmTemp ) != 0; i++ )
+		{
+			tempResX.push_back( dmTemp.dmPelsWidth );
+			tempResY.push_back( dmTemp.dmPelsHeight );
+		}
+
+		maxResX = *std::max_element( tempResX.begin( ), tempResX.end( ) );
+		maxResY = *std::max_element( tempResY.begin( ), tempResY.end( ) );
+
+		dmTemp.dmPelsWidth = maxResX;
+		dmTemp.dmPelsHeight = maxResY;
+		dmTemp.dmDisplayFrequency = m_dm.dmDisplayFrequency;
+
+		if( ChangeDisplaySettings( &dmTemp, CDS_FULLSCREEN ) != DISP_CHANGE_SUCCESSFUL )
+		{
+			SwitchResolution( true );
+			MessageBox( 0, L"Display resolution switch failure", L"ERROR: Display resolution switch failure", MB_ICONERROR | MB_OK );
+		}
+		Sleep( 2500 );
+	}
+	else
+	{
+		if( ChangeDisplaySettings( &m_dm, CDS_FULLSCREEN ) != DISP_CHANGE_SUCCESSFUL )
+		{
+			MessageBox( 0, L"Display resolution switch failure", L"ERROR: Display resolution switch failure", MB_ICONERROR | MB_OK );
+		}	
+		Sleep( 2500 );
+	}
+}
+
 bool System::Init( )
 {
 	if( GetAppId( ) )
@@ -283,8 +428,15 @@ void System::Shutdown( )
 		TerminateProcess( m_hExplorer, 0 );
 		CloseHandle( m_hExplorer );
 	}
-	TerminateProcess( m_hProcess, 0 );
-	CloseHandle( m_hProcess );
+	if( m_hProcess != nullptr )
+	{
+		TerminateProcess( m_hProcess, 0 );
+		CloseHandle( m_hProcess );
+	}
+	if( m_bUseDSR )
+	{
+		SwitchResolution( true );
+	}
 	CoUninitialize( );
 }
